mcp-stocks/src/requirements.txt
mcp==1.8.1
mcp[cli]==1.8.1
python-dotenv==1.1.0
requests==2.32.3
uvicorn==0.34.2
yfinance==0.2.61